Adicionadas funções: matraiz, matsin, matcos, mattan, matasin, matacos, matatan, matexp, matlog e arqsav.limpou. Correções para funcionar em 64 bits (foi feito para funcionar, mas nunca foi testado) Correção de bug: uso indevido de memcpy quando deveria ser memmove. Correção na classe MUDAPROG, no código interpretado. Manual revisado até o tópico 15.2. No MUD: uso da função matexp ao invés de usar uma tabela. No MUD: pequenas mudanças quando o jogador sai do MUD mas o personagem fica.

+++ socket.cpp
@@ -145,7 +145,7 @@ TSocket * TSocket::Conectar(const char * ender, int porta, bool ssl)
     conSock.sin_family=AF_INET;
     conSock.sin_port=htons(porta);
     conSock.sin_addr.s_addr=inet_addr(ender);
-    if ( (conSock.sin_addr.s_addr) == (unsigned long)-1 )
+    if ( (conSock.sin_addr.s_addr) == (unsigned int)-1 )
     {
         if ( (hnome=gethostbyname(ender)) == NULL )
             return 0;
